requirements.txt
absl-py==2.1.0
argcomplete==3.1.2
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
debugpy==1.7.0
decorator==4.4.2
deepdiff==5.0.1
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
enum34==1.1.10
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
higher==0.2.1
idna==2.10
imageio==2.8.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.4.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonpickle==1.4.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-server==1.0.8
jupyter_nbextensions_configurator==0.6.4
jupyterlab-pygments==0.2.2
kiwisolver==1.2.0
llvmlite==0.34.0
lxml==5.3.0
Markdown==3.2.2
MarkupSafe==2.1.5
matplotlib==3.2.2
matplotlib-inline==0.1.6
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
mpyq==0.2.5
munch==2.5.0
nbclient==0.5.11
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.6.0
notebook==6.0.3
numba==0.51.2
numpy==1.19.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
opencv-python==4.2.0.34
ordered-set==4.0.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathlib2==2.3.5
pbr==5.4.5
pexpect==4.9.0
pickleshare==0.7.5
Pillow==7.2.0
pip-review==1.1.0
pkgutil_resolve_name==1.3.10
pluggy==0.13.1
portpicker==1.3.1
probscale==0.2.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py==1.9.0
py-cpuinfo==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclustering==0.10.1.2
pygame==2.0.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySC2==3.0.0
pytest==5.4.3
python-dateutil==2.8.1
python-ternary==1.0.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==5.2.2
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s2clientprotocol==4.11.4.78285.0
s2protocol==4.11.4.78285.0
sacred==0.8.1
scipy==1.5.0
Send2Trash==1.5.0
six==1.15.0
sk-video==1.1.10
smmap==3.0.4
snakeviz==2.1.0
tensorboard==2.2.2
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
torch==1.13.1
tornado==6.0.4
traitlets==4.3.3
typing_extensions==4.7.1
urllib3==1.25.9
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
